+++ 2026spring/frontend/src/app/derivative/streams/page.tsx
@@ -1,12 +1,12 @@
 import { Suspense } from "react";
 import StreamsContent from "./StreamsContent";
 import { getDerivativeStreams } from "@/lib/derivative";
-import { getDerivativeArchive } from "@/lib/derivative";
+import { getDerivativeArchives } from "@/lib/derivative";
 
 export default async function Page() {
   const [schedule, archive] = await Promise.all([
     getDerivativeStreams(),
-    getDerivativeArchive(),
+    getDerivativeArchives(),
   ]);
 
   return (
